requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
PySide6>=6.6.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
Pillow>=10.0.0
pytest>=8.2.0
requests>=2.32.0